Don't cast int64 attributes to int32
Summary:
Attribute map takes in int64 and stores as int64 so we
don't need to cast it to int32 here.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Maniphest Tasks: T24504

Differential Revision: https://phabricator.sourcevertex.net/D44511

+++ willow/src/builder_impl.cpp
@@ -790,10 +790,10 @@ void BuilderImpl::addNodeAttribute(const std::string &attributeName,
     attr.set_i(popart::any_cast<uint32_t>(attributeValue));
   } else if (tinfo == typeid(int64_t)) {
     attr.set_type(ONNX_NAMESPACE::AttributeProto::INT);
-    attr.set_i(static_cast<int>(popart::any_cast<int64_t>(attributeValue)));
+    attr.set_i(popart::any_cast<int64_t>(attributeValue));
   } else if (tinfo == typeid(uint64_t)) {
     attr.set_type(ONNX_NAMESPACE::AttributeProto::INT);
-    attr.set_i(static_cast<int>(popart::any_cast<uint64_t>(attributeValue)));
+    attr.set_i(popart::any_cast<uint64_t>(attributeValue));
   } else if (tinfo == typeid(std::vector<int64_t>)) {
     attr.set_type(ONNX_NAMESPACE::AttributeProto::INTS);
     const std::vector<int64_t> &values =
